Move runCatchingExceptCancel to datasource
